[update] 소스 분리 및 위치 조정
